README.md  description
The currency converter is my second project while learning my profession as a frontend developer. It consists in converting Polish zlotys into other currencies. Easy to use because you just need to select the currency and enter the amount in PLN and after clicking the button we will get the value of the currency.